api/requirements.txt
Flask==1.1.2
Keras==2.3.1
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
tensorflow==1.13.1
gunicorn==20.0.4
gevent